+++ lib/include/types.h
@@ -3,7 +3,7 @@
 #include "platform.h"
 #include <cstdint>
 
-typedef int32_t key_type;
+typedef size_t key_type;
 
 class Camera;
 class Planet;
